refactoring & added friend service

+++ ChessAPIServer/Models/RedisDB.cs
@@ -4,7 +4,7 @@ namespace APIServer.Models;
 // redisdb => rdb
 public class RdbAuthUserData
 {
-    public int Uid { get; set; } = 0;
+    public long Uid { get; set; } = 0;
     public string Token { get; set; } = "";
 }
